Refactor CRSF telemetry
This commit introduces Custom CRSF Telemetry.

It is using a "custom" CRSF frame (0x88) for passing the telemetry
data to the Tx. On the Tx side, a LUA script is parsing these frames
and injecting the sensors into the native sensor framework.

The FC has over hundred telemetry sensors to choose from.
Up to 40 can be enabled at the same time. Depending on the link
bandwidth (up to 20kbit/s with ELRS), sensor update rates up to 50Hz
are achievable.

This framework is also fully extensible in the future. For adding
new sensors, changes in the firmware and the LUA script is needed.
The Receiver or the Transmitter are not involved in the process.