Add support for FrSky FBUS
This commit adds two receiver protocols, FBUS and FPORT2.

F.Port2 is an older version of FBUS that was running on a lower rate.
There are still receivers in the market that are advertised as
FBUS compatible, but are actually running the slower version of it.

Please use FPORT2 protocol with such receivers, for example R-XSR.